+++ .gitignore
@@ -1,8 +1,10 @@
-#ignore everything
-*
+#ignore python cache
+*.pyc
 
-#but not these files
-!*.py
-!.gitignore
-!/Gui
+#ignore txt
+*.txt
+
+#ignore folders
+/__pycache__
+Gui/__pycache__
