+++ Gui/Icon.py
@@ -4,4 +4,3 @@ class Icon(Gtk.Image):
 	def __init__(self, ico_name):
 		ico = Gio.ThemedIcon(name=ico_name)
 		Gtk.Image.__init__(self, gicon=ico, icon_size=Gtk.IconSize.BUTTON)
-		#self.new_from_gicon(ico, Gtk.IconSize.BUTTON)
